Abstract away createNftBuyOffer
